test(mcp): pin the retry re-publish path, harden attempts against wraparound
Self-review pass on the #2513 teardown-retry design: the retry re-entry
branch (a retry whose Step-1 frame build failed on attempt 1, then
succeeds on retry) had zero coverage - the design doc specced it, the
implementation skipped it. Adds a test forcing that exact path (pressure
victim + one-shot terminal-build fault + one-shot unsubscribe fault) and
asserts the retry actually re-publishes (ring frame count), not that it
silently treats the record as already handled - the #2517-shaped failure
mode this would reproduce under a new name. Verified by temporarily
reintroducing that exact bug and confirming the new assertion goes red.

Also: widens BridgeRecord::teardown_attempts from uint8_t to size_t (it's
compared directly against the size_t, test-settable teardown_retry_max;
a narrower counter wrapping at 256 would silently re-open retry
eligibility forever), and softens two "retry_max = 0 restores the
pre-#2513 one-way behaviour" claims to say what's actually true: the
RETENTION posture reverts, but the exhausted-outcome counter/log/audit
row are new observability that still fires once even at 0.

+++ server/core/src/mcp_stream_bridge.hpp
@@ -266,8 +266,12 @@ public:
         /// attempt has had multiple ticks to self-heal and is effectively
         /// permanent; retrying it forever would re-run the same audit/log/metric
         /// work every tick for a record that is never coming back. `0` restores
-        /// the pre-#2513 one-way behaviour (retained until shutdown, no retry) -
-        /// a code-constant default only, deliberately not exposed as a CLI
+        /// the pre-#2513 one-way RETENTION posture (still retained until
+        /// shutdown, never retried) - it does NOT restore pre-#2513
+        /// observability: the first bail still fires `teardown_retry_total{
+        /// outcome="exhausted"}`, the exhaustion error log, and the "(retry
+        /// budget exhausted...)" audit suffix, all of which are new. A
+        /// code-constant default only, deliberately not exposed as a CLI
         /// flag/env var: this is a fault-recovery bound, not an operator dial.
         std::size_t teardown_retry_max = 3;
     };
@@ -994,8 +998,13 @@ private:
         /// touching torn_down's own claim sites.
         bool teardown_retry_claimable = false;
         /// Attempts consumed so far; 0 before the first entry to teardown_claimed,
-        /// bounded by Config::teardown_retry_max beyond the first.
-        std::uint8_t teardown_attempts = 0;
+        /// bounded by Config::teardown_retry_max beyond the first. std::size_t,
+        /// not a narrower counter: it is compared directly against
+        /// Config::teardown_retry_max (also std::size_t, test-settable), and a
+        /// narrower type wrapping at its max would silently re-open eligibility
+        /// forever - the exact unbounded-retry outcome this bound exists to
+        /// prevent.
+        std::size_t teardown_attempts = 0;
         /// The `decision` teardown_claimed's Step 1 ran (or will run) with,
         /// persisted so a retry pass can replay the SAME decision without
         /// independently re-arbitrating what to publish.
